allow to hide embedded parsing using feature flag
