+++ api/feature/tests/test_embedded_media.py
@@ -1,16 +1,16 @@
 from django.conf import settings
-from django.test import TestCase, override_settings
-from rest_framework.test import APIClient
+from django.test import override_settings
+from rest_framework.test import APIClient, APITestCase
 
-from api.models import User
-from api.models.file import (
+from api.feature.embedded_media import (
     GOOGLE_PIXEL_MOTION_PHOTO_MP4_SIGNATURES,
     JPEG_EOI_MARKER,
     SAMSUNG_MOTION_PHOTO_MARKER,
-    File,
     extract_embedded_media,
     has_embedded_media,
 )
+from api.models import User
+from api.models.file import File
 from api.tests.utils import create_test_photo, create_test_user
 
 
@@ -29,7 +29,7 @@ RANDOM_BYTES = b"\x13\x37\xc0\xde"
 
 
 @override_settings(MEDIA_ROOT="/tmp")
-class MotionPhotoTest(TestCase):
+class EmbeddedMediaTest(APITestCase):
     def setUp(self):
         self.test_image_path = "/tmp/test_file.jpeg"
         self.test_video_path = "/tmp/test_file.mp4"
@@ -38,32 +38,32 @@ class MotionPhotoTest(TestCase):
 
     def test_should_not_process_non_jpeg_files(self):
         file = create_test_file(self.test_video_path, self.user, MP4)
-        actual = has_embedded_media(file)
+        actual = has_embedded_media(file.path)
         self.assertFalse(actual)
 
     def test_google_pixel_motion_photo_signatures(self):
         for signature in GOOGLE_PIXEL_MOTION_PHOTO_MP4_SIGNATURES:
             content = JPEG + MP4_PREFIX + signature + MP4_DATA
             file = create_test_file(self.test_image_path, self.user, content)
-            actual = has_embedded_media(file)
+            actual = has_embedded_media(file.path)
             self.assertTrue(actual)
 
     def test_samsung_motion_photo_signature(self):
         content = JPEG + SAMSUNG_MOTION_PHOTO_MARKER + MP4_DATA
         file = create_test_file(self.test_image_path, self.user, content)
-        actual = has_embedded_media(file)
+        actual = has_embedded_media(file.path)
         self.assertTrue(actual)
 
     def test_other_content_should_not_report_as_having_embedded_media(self):
         file = create_test_file(self.test_image_path, self.user, RANDOM_BYTES)
-        actual = has_embedded_media(file)
+        actual = has_embedded_media(file.path)
         self.assertFalse(actual)
 
     def test_extract_embedded_media_from_google_motion_photo(self):
         for signature in GOOGLE_PIXEL_MOTION_PHOTO_MP4_SIGNATURES:
             content = JPEG + MP4_PREFIX + signature + MP4_DATA
             file = create_test_file(self.test_image_path, self.user, content)
-            path = extract_embedded_media(file)
+            path = extract_embedded_media(file.path, file.hash)
             expected = f"{settings.MEDIA_ROOT}/embedded_media/{file.hash}_1.mp4"
             self.assertEqual(path, expected)
             with open(path, "rb") as f:
@@ -73,7 +73,7 @@ class MotionPhotoTest(TestCase):
     def test_extract_embedded_media_from_samsung_motion_photo(self):
         content = JPEG + SAMSUNG_MOTION_PHOTO_MARKER + MP4
         file = create_test_file(self.test_image_path, self.user, content)
-        path = extract_embedded_media(file)
+        path = extract_embedded_media(file.path, file.hash)
         expected = f"{settings.MEDIA_ROOT}/embedded_media/{file.hash}_1.mp4"
         self.assertEqual(expected, path)
         with open(path, "rb+") as f:
