+++ librephotos/settings/test.py
@@ -1,3 +1,16 @@
 from .production import *  # noqa
 
 DEBUG = True
+LOGGING = {
+    "version": 1,
+    "disable_existing_loggers": True,
+    "handlers": {
+        "null": {
+            "class": "logging.NullHandler",
+        },
+    },
+    "root": {
+        "handlers": ["null"],
+        "level": "CRITICAL",
+    },
+}
